requirements.txt
pydantic
fastapi
python-dotenv
uvicorn